bring back old pkg/config, move new stuff to pkg/config_store
