+++ pkg/volumes/volumes.go
@@ -1,7 +1,7 @@
 package volumes
 
 import (
-	"github.com/Yelp/paasta-tools-go/pkg/config"
+	"github.com/Yelp/paasta-tools-go/pkg/config_store"
 )
 
 type VolumeConfig struct {
@@ -14,7 +14,9 @@ type Volume struct {
 	Mode          string `json:"mode" mapstructure:"mode"`
 }
 
-func DefaultVolumesFromReader(configStore *config.Store) (volumes []Volume, err error) {
+func DefaultVolumesFromReader(
+	configStore *config_store.Store,
+) (volumes []Volume, err error) {
 	volumeConfig := &VolumeConfig{Volumes: []Volume{}}
 	err = configStore.Load("volumes", &volumeConfig.Volumes)
 	return volumeConfig.Volumes, err
